requirements.txt
requests==2.32.3
openai==1.57.0
openai 
tiktoken
PynaCL
customtkinter
keyboard
google-generativeai
pillow